requirements.txt
# Driving License Detection - ML Training Pipeline
# ================================================

# Core ML frameworks
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Computer vision
opencv-python>=4.8.0
albumentations>=1.3.0

# Data and utilities
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
Pillow>=9.5.0

# Visualization and logging
matplotlib>=3.7.0
tensorboard>=2.13.0
seaborn>=0.12.0

# Optional: MIDV-500 dataset download
# midv500>=0.2.1

# WebSocket server (like driver_safety_system)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Qwen vision-language model for OCR + validation (local, no external API)
transformers>=4.39.0
accelerate>=0.28.0
qwen-vl-utils>=0.0.8
